add a cleanup script
The code style for this driver does not conform to upstream Linux kernel
style expectations. Having this driver included in mainline Linux
kernels is really desirable, so this commit adds a script to reformat
the code, as a starting point to prepare the code for the merging
process.

The intention is that the script will be extended and used on future
vendor code drops to convert to something that's closer to mergeable.

Secondly, any patches should not be applied directly on top of the
vendor code, but on top of the Linux code style, so that said patches
can also be applied to future vendor code drops, even after the vendor
has made large formatting changes.