remove redundant return from void functions
This applies the same transform as from the
topic/remove-redundant-return-void branch.

The 3.6.14.1 and 3.6.14.3 branches have diverged enough that we can't do
simple merges of topic branches anymore, but we can still apply the same
transform by running the cleanup.sh script.

That proves the point of scripting the cleanup changes!