Fix building of MV88X3310 with DKMS
Signed-off-by: Jonathan Bennett <JBennett@incomsystems.biz>
I have a card that uses the MV88X3310 chip, and have located the appropriate firmware file and added it to my source directory. DKMS was failing to find and process that firmware file, because the dkms build instructions were changing the current directory before the Makefile had a chance to test for the hdr file's existence. This is a potential workaround, tested to work on CentOS 8.