use more specific logging macros when available
This change ensures that all driver logging output is attributable to
the driver and device.

The order of preference for logging macros is:
netdev_<level> when a struct netdev is available,
dev_<level> when a struct netdev is not available but struct device is,
pr_<level> when a struct device is not available.

Also define pr_fmt to include the module name.