remove driver state tracking from tn40xx
The tn40xx driver is unusual in the sense that it tries to keep its own
state through the start/stop/suspend lifecycle. It's redundant, because
the device driver model is designed such that these state transitions
are known at driver development time.

The driver only needs to track the hardware state and ensure that the
netif_* and netdev_* APIs are called correctly. Adding additional state
tracking in the driver simply obscures bugs.

Signed-off-by: Andrew Cooks <acooks@gmail.com>